Add isolation to RemoteCI and Jobs.
A user can only do CRUD operations on the remotecis
and jobs of its team.

+++ dci/server/api/v1/jobs.py
@@ -47,6 +47,11 @@ def _verify_existence_and_get_job(job_id):
 @auth2.requires_auth
 def create_jobs(user_info):
     values = schemas.job.post(flask.request.json)
+
+    # If it's not a super admin nor belongs to the same team_id
+    v1_utils.check_user_is_super_admin_or_same_team(user_info,
+                                                    values.get('team_id'))
+
     etag = utils.gen_etag()
     values.update(
         {'id': utils.gen_uuid(),
@@ -85,6 +90,10 @@ def get_all_jobs(user_info, jd_id=None):
     if embed:
         query = v1_utils.get_query_with_join(models.JOBS, [models.JOBS],
                                              embed, _VALID_EMBED)
+
+    user_info_team_id = user_info.get('team_id')
+    query = v1_utils.add_filter_if_not_super_admin(
+        query, user_info, models.JOBS.c.team_id == user_info_team_id)
 
     query = v1_utils.sort_query(query, args['sort'], _JOBS_COLUMNS)
     query = v1_utils.where_query(query, args['where'], models.JOBS,
@@ -132,6 +141,10 @@ def get_job_by_id(user_info, jd_id):
         query = v1_utils.get_query_with_join(models.JOBS, [models.JOBS],
                                              embed, _VALID_EMBED)
 
+    user_info_team_id = user_info.get('team_id')
+    query = v1_utils.add_filter_if_not_super_admin(
+        query, user_info, models.JOBS.c.team_id == user_info_team_id)
+
     query = query.where(models.JOBS.c.id == jd_id)
 
     row = flask.g.db_conn.execute(query).fetchone()
@@ -153,7 +166,10 @@ def delete_job_by_id(user_info, jd_id):
     # get If-Match header
     if_match_etag = utils.check_and_get_etag(flask.request.headers)
 
-    _verify_existence_and_get_job(jd_id)
+    job = dict(_verify_existence_and_get_job(jd_id))
+
+    v1_utils.check_user_is_super_admin_or_same_team(user_info,
+                                                    job['team_id'])
 
     query = models.JOBS.delete().where(
         sqlalchemy.sql.and_(models.JOBS.c.id == jd_id,
